Crate cargo_util

Miscellaneous support code used by Cargo.

This crate is maintained by the Cargo team, primarily for use by Cargo and not intended for external use (except as a transitive dependency). This crate may make major changes to its APIs or be deprecated without warning.

Modules

du 🔒

A simple disk usage estimator.

paths

Various utilities for working with files and paths.

process_builder 🔒

process_error 🔒

Error value for crate::ProcessBuilder when a process fails.

read2 🔒

registry

sha256 🔒

Structs

ProcessBuilder

A builder object for an external process, similar to std::process::Command.

ProcessError

Sha256

Functions

du

Determines the disk usage of all files in the given directory.

exit_status_to_string

Converts an ExitStatus to a human-readable string suitable for displaying to a user.

is_ci

Whether or not this running in a Continuous Integration environment.

is_simple_exit_code

Returns true if the given process exit code is something a normal process would exit with.

read2