cargo_util::paths

Function strip_prefix_canonical

pub fn strip_prefix_canonical(
    path: impl AsRef<Path>,
    base: impl AsRef<Path>,
) -> Result<PathBuf, StripPrefixError>

Strips base from path.

This canonicalizes both paths before stripping. This is useful if the paths are obtained in different ways, and one or the other may or may not have been normalized in some way.