cargo_util::process_error
Struct ProcessError

pub struct ProcessError {
    pub desc: String,
    pub code: Option<i32>,
    pub stdout: Option<Vec<u8>>,
    pub stderr: Option<Vec<u8>>,
}

Fields
desc: String

A detailed description to show to the user why the process failed.

code: Option<i32>

The exit status of the process.

This can be None if the process failed to launch (like process not found) or if the exit status wasnโt a code but was instead something like termination via a signal.

stdout: Option<Vec<u8>>

The stdout from the process.

This can be None if the process failed to launch, or the output was not captured.

stderr: Option<Vec<u8>>

The stderr from the process.

This can be None if the process failed to launch, or the output was not captured.

Implementations

impl ProcessError

pub fn new( msg: &str, status: Option<ExitStatus>, output: Option<&Output>, ) -> ProcessError

Creates a new ProcessError.

status can be None if the process did not launch.

output can be None if the process did not launch, or output was not captured.

pub fn new_raw( msg: &str, code: Option<i32>, status: &str, stdout: Option<&[u8]>, stderr: Option<&[u8]>, ) -> ProcessError

Creates a new ProcessError with the raw output data.

code can be None for situations like being killed by a signal on unix.

pub fn could_not_execute(cmd: impl Display) -> ProcessError

Creates a ProcessError with โcould not execute process {cmd}โ.

cmd is usually but not limited to std::process::Command.