cargo_util
Struct ProcessBuilder

pub struct ProcessBuilder {
    program: OsString,
    args: Vec<OsString>,
    env: BTreeMap<String, Option<OsString>>,
    cwd: Option<OsString>,
    wrappers: Vec<OsString>,
    jobserver: Option<Client>,
    display_env_vars: bool,
    retry_with_argfile: bool,
    stdin: Option<Vec<u8>>,
}

A builder object for an external process, similar to std::process::Command.

Fields
program: OsString

The program to execute.

args: Vec<OsString>

A list of arguments to pass to the program.

env: BTreeMap<String, Option<OsString>>

Any environment variables that should be set for the program.

cwd: Option<OsString>

The directory to run the program from.

wrappers: Vec<OsString>

A list of wrappers that wrap the original program when calling ProcessBuilder::wrapped. The last one is the outermost one.

jobserver: Option<Client>

The make jobserver. See the jobserver crate for more information.

display_env_vars: bool

true to include environment variable in display.

retry_with_argfile: bool

true to retry with an argfile if hitting “command line too big” error. See ProcessBuilder::retry_with_argfile for more information.

stdin: Option<Vec<u8>>

Data to write to stdin.

Implementations

impl ProcessBuilder

pub fn new<T: AsRef<OsStr>>(cmd: T) -> ProcessBuilder

Creates a new ProcessBuilder with the given executable path.

pub fn program<T: AsRef<OsStr>>(&mut self, program: T) -> &mut ProcessBuilder

(chainable) Sets the executable for the process.

pub fn arg<T: AsRef<OsStr>>(&mut self, arg: T) -> &mut ProcessBuilder

(chainable) Adds arg to the args list.

pub fn args<T: AsRef<OsStr>>(&mut self, args: &[T]) -> &mut ProcessBuilder

(chainable) Adds multiple args to the args list.

pub fn args_replace<T: AsRef<OsStr>>( &mut self, args: &[T], ) -> &mut ProcessBuilder

(chainable) Replaces the args list with the given args.

pub fn cwd<T: AsRef<OsStr>>(&mut self, path: T) -> &mut ProcessBuilder

(chainable) Sets the current working directory of the process.

pub fn env<T: AsRef<OsStr>>(&mut self, key: &str, val: T) -> &mut ProcessBuilder

(chainable) Sets an environment variable for the process.

pub fn env_remove(&mut self, key: &str) -> &mut ProcessBuilder

(chainable) Unsets an environment variable for the process.

pub fn get_program(&self) -> &OsString

Gets the executable name.

pub fn get_args(&self) -> impl Iterator<Item = &OsString>

Gets the program arguments.

pub fn get_cwd(&self) -> Option<&Path>

Gets the current working directory for the process.

pub fn get_env(&self, var: &str) -> Option<OsString>

Gets an environment variable as the process will see it (will inherit from environment unless explicitally unset).

pub fn get_envs(&self) -> &BTreeMap<String, Option<OsString>>

Gets all environment variables explicitly set or unset for the process (not inherited vars).

pub fn inherit_jobserver(&mut self, jobserver: &Client) -> &mut Self

Sets the make jobserver. See the jobserver crate for more information.

pub fn display_env_vars(&mut self) -> &mut Self

Enables environment variable display.

pub fn retry_with_argfile(&mut self, enabled: bool) -> &mut Self

Enables retrying with an argfile if hitting “command line too big” error

This is primarily for the @path arg of rustc and rustdoc, which treat each line as an command-line argument, so LF and CRLF bytes are not valid as an argument for argfile at this moment. For example, RUSTDOCFLAGS="--crate-version foo\nbar" cargo doc is valid when invoking from command-line but not from argfile.

To sum up, the limitations of the argfile are:

Must be valid UTF-8 encoded.

Must not contain any newlines in each argument.

Ref:

https://doc.rust-lang.org/rustdoc/command-line-arguments.html#path-load-command-line-flags-from-a-path

https://doc.rust-lang.org/rustc/command-line-arguments.html#path-load-command-line-flags-from-a-path

pub fn stdin<T: Into<Vec<u8>>>(&mut self, stdin: T) -> &mut Self

Sets a value that will be written to stdin of the process on launch.

fn should_retry_with_argfile(&self, err: &Error) -> bool

pub fn status(&self) -> Result<ExitStatus>

Like Command::status but with a better error message.

fn _status(&self) -> Result<ExitStatus>

pub fn exec(&self) -> Result<()>

Runs the process, waiting for completion, and mapping non-success exit codes to an error.

pub fn exec_replace(&self) -> Result<()>

Replaces the current process with the target process.

On Unix, this executes the process using the Unix syscall execvp, which will block this process, and will only return if there is an error.

On Windows this isn’t technically possible. Instead we emulate it to the best of our ability. One aspect we fix here is that we specify a handler for the Ctrl-C handler. In doing so (and by effectively ignoring it) we should emulate proxying Ctrl-C handling to the application at hand, which will either terminate or handle it itself. According to Microsoft’s documentation at https://docs.microsoft.com/en-us/windows/console/ctrl-c-and-ctrl-break-signals. the Ctrl-C signal is sent to all processes attached to a terminal, which should include our child process. If the child terminates then we’ll reap them in Cargo pretty quickly, and if the child handles the signal then we won’t terminate (and we shouldn’t!) until the process itself later exits.

pub fn output(&self) -> Result<Output>

Like Command::output but with a better error message.

fn _output(&self) -> Result<Output>

pub fn exec_with_output(&self) -> Result<Output>

Executes the process, returning the stdio output, or an error if non-zero exit status.

pub fn exec_with_streaming( &self, on_stdout_line: &mut dyn FnMut(&str) -> Result<()>, on_stderr_line: &mut dyn FnMut(&str) -> Result<()>, capture_output: bool, ) -> Result<Output>

Executes a command, passing each line of stdout and stderr to the supplied callbacks, which can mutate the string data.

If any invocations of these function return an error, it will be propagated.

If capture_output is true, then all the output will also be buffered and stored in the returned Output object. If it is false, no caching is done, and the callbacks are solely responsible for handling the output.

fn build_command_with_argfile(&self) -> Result<(Command, NamedTempFile)>

Builds the command with an @<path> argfile that contains all the arguments. This is primarily served for rustc/rustdoc command family.

fn build_command_without_args(&self) -> Command

Builds a command from ProcessBuilder for everything but not args.

pub fn build_command(&self) -> Command

Converts ProcessBuilder into a std::process::Command, and handles the jobserver, if present.

Note that this method doesn’t take argfile fallback into account. The caller should handle it by themselves.

pub fn wrapped(self, wrapper: Option<impl AsRef<OsStr>>) -> Self

Wraps an existing command with the provided wrapper, if it is present and valid.

Examples

use cargo_util::ProcessBuilder;
// Running this would execute `rustc`
let cmd = ProcessBuilder::new("rustc");

// Running this will execute `sccache rustc`
let cmd = cmd.wrapped(Some("sccache"));